merge(upstream): memory-error paths, posix_spawn and atomic logging from DeusData/codebase-memory-mcp@80afcd6b,579a9108,7814dacf,c51482ae,9a41d831
80afcd6b — the clang-analyzer lane's real findings, ported to the sites that
exist here. LEAKS: mcp get_architecture leaked scope_path on the missing-store
early return (REQUIRE_STORE frees only `project`) — allocate after the gate;
pass_definitions leaked the whole pass-owned result cache when cancellation hit
mid-extraction; the store package-boundary row-scan abort path freed the node
arrays but not the boundary accumulators or their duplicated package strings;
the cbm quarantine set leaked the replaced value on a duplicate path line (the
table borrows key pointers, so a fresh key copy leaked too) and leaked the
surviving half on a partial strdup failure; pass_githistory's unchecked
malloc/strdup dereferenced NULL on OOM and leaked the index cell on a failed
strdup — allocate before claiming the slot. NULL/UB: the cli config subcommand
took NULL argv with nonzero argc straight into argv[0]; store bfs_multi with a
negative max_results freed fields of an unwritten negative-index slot;
pass_calls emit_http_async_edge could deref a NULL target if the two
hand-duplicated URL predicates ever drifted; both sqlite_writer leaf-array OOM
paths left leaf_count stale beside a NULL array, walking pb_finalize_* into
leaves[0]. Plus the hardening the commit describes (Leiden CSR + aggregate
arrays and SCC adjacency to calloc + endpoint guards, the ncyc==0 no-slot
invariant made local).

579a9108 — fork() duplicates the parent's whole address-space bookkeeping, and
under an ASan-instrumented parent that includes an enormous shadow mapping;
past a footprint threshold the child is jetsam-killed BEFORE exec replaces the
image, so the reaped corpse reads as the launched tool dying (a fixture's
`git init` "failing" only once enough suites have run ahead of it). macOS now
uses posix_spawn, which never copies the parent address space.

7814dacf (zstd half only) — decompression already took size_t but compression
still took int for both the source length and the destination capacity, so a
database past 2 GiB handed the encoder a negative length. Both lengths and the
bound helper are size_t now and the function returns int64_t like its
decompressing counterpart. The artifact.cpp call site is updated in lockstep:
without it the fix is cosmetic, and worse, `(int)bound` on a huge bound goes
negative and widens back to an enormous size_t capacity.

c51482ae + 9a41d831 as one end state — logging configuration is atomic and
emit_line loads the sink ONCE into a local, closing the test-then-call window.

Fork-side adaptations:

- 579a9108 could not be applied as a diff: our subprocess.cpp is at a much
  earlier stage than upstream's (a single synchronous cbm_run_posix with an
  inline fork+exec child; none of upstream's async spawn/poll layer,
  cbm_posix_child_exec, setpgid, fd-close loop or kill-tree). Re-implemented
  against our shape, preserving OUR path's guarantees exactly. Two upstream
  attributes are deliberately NOT set:
    * POSIX_SPAWN_SETPGROUP — this fork has no kill-tree at all; the only kill
      is kill(pid, SIGKILL) on quiet-timeout. A new process group on macOS only
      would stop the worker receiving terminal SIGINT, so Ctrl-C would orphan
      an index worker on macOS while Linux still killed it — a macOS-only
      regression with no compensating group-kill.
    * POSIX_SPAWN_CLOEXEC_DEFAULT — our child has no close-everything loop and
      does not rebind stdin, so this would close inherited fd 0 (the MCP
      transport). Upstream can use it only because its child dup2s /dev/null
      onto stdin.
  An exec-class errno falls back to fork+exec so a bogus binary still yields
  the documented "child exits 127" classification rather than our -1
  SPAWN_FAILED, which the supervisor reads as "degrade in process".
  _NSGetEnviron() replaces upstream's `extern char **environ;` — macOS does not
  declare environ, and a bare extern in a C++ TU would take C++ linkage.
- c51482ae uses std::atomic, not <stdatomic.h>/_Atomic (not a C++ keyword).
  9a41d831 then folds in for free: std::atomic's constructor is constexpr, so
  the initializer is a constant by construction and upstream's
  (cbm_log_sink_fn)NULL cast is unnecessary. No foundation compat header was
  added, keeping the MinGW <windows.h> macro hazard away from this TU.
  Partly inert: this fork has no CBMLogSinkMode/cbm_log_set_sink_ex, so the
  store-mode-before-sink ordering hunk has nothing to apply to.
- 80afcd6b's cli hunk restructures a --help arm added by the unported daemon
  work; only the substantive half (`argc == 0 || !argv`) applies, and the NULL
  deref was genuinely reachable here.

Inert, not ported: 80afcd6b's src/cli/config_yaml_edit.c hunk (the file does
not exist in this fork) and its read_bounded_line memset hunk (our
Content-Length transport uses cbm_getline, not upstream's growable reader).

Also fixed, found while porting: tests/test_zstd.cpp carried hand-rolled
extern "C" prototypes that had already DRIFTED from the real API — it declared
cbm_zstd_decompress as int(const char*, int, char*, int) against a real
int64_t(const char*, size_t, char*, size_t). That is a live ABI mismatch
predating this sync. Replaced with the real header.

Tests: +2. zstd_compress_bound_past_int_max_does_not_wrap is RED-on-revert
WITHOUT needing a 2 GiB buffer (with the old int parameter the input truncates
to INT_MIN, whose size_t reinterpretation exceeds ZSTD_MAX_INPUT_SIZE, so
ZSTD_compressBound returns 0), plus a static_assert pinning the return width.
log_concurrent_sink_swap_is_defined is a TSan tripwire (3 logger threads x 500
lines against a sink/format flipper) — upstream shipped no test for c51482ae,
and this is honestly not RED-on-revert outside TSan, since an aligned pointer
rarely tears in practice.

Verified: full suite 5615 passed / 0 failed (non-sanitizer Apple clang);
clang-format-20.1.8, cppcheck and the NOLINT gate clean over every changed TU.

+++ src/pipeline/artifact.cpp
@@ -541,15 +541,17 @@ int cbm_artifact_export(const char *db_path, const char *repo_path, const char *
         return artifact_export_fail("read_db", db_path, "empty_or_unreadable", errno);
     }
 
-    /* Compress with zstd */
-    size_t bound = cbm_zstd_compress_bound((int)db_size);
+    /* Compress with zstd. Sizes stay size_t end-to-end: casting db_size down to
+     * int handed the encoder a negative length past 2 GiB, and (int)bound on a
+     * huge bound went negative and widened back to an enormous capacity. */
+    size_t bound = cbm_zstd_compress_bound(db_size);
     char *compressed = (char *)malloc(bound);
     if (!compressed) {
         free(db_data);
         return artifact_export_fail("compress", NULL, "alloc_compressed_buffer", 0);
     }
 
-    int clen = cbm_zstd_compress(db_data, (int)db_size, compressed, (int)bound, compression_level);
+    int64_t clen = cbm_zstd_compress(db_data, db_size, compressed, bound, compression_level);
     free(db_data);
 
     if (clen <= 0) {
